FileUpload/requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.8
et_xmlfile==2.0.0
inflection==0.5.1
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
